possible last update
theres no snow, no sounds for the radio, and no models of actual characters. so this project has gone great. w/ out the radio or snow i think the main theme of the game is just non-existant, i might try and get the snow in
probs the last update ive got time for before submission.